Remove lib.rs and add issue.rs
